Strip accelerator keys from certain menu items (#15596)
Fixes #15588
Replaces #15589
Follow-up of #15364

Summary of the issue:
Among changes made in #15364, the accelerator key for the item "Reload plugins" in "Tools" menu has be modified to "E".
Users of NVDA remote add-on have complained that it conflicts with this add-on's accelerator key, that is being used for many years.

Usually, NVDA defines the shortcut keys and the add-ons should adapt to it. Given the popularity of NVDA Remote and it's age in the community though, NV Access may accept this case as a valid exception to this rule.

Description of user facing changes
Strip accelerator keys from "reload plugins" and "run com registration fixing tool".

Description of development approach
When removing the accelerator key for reload plugins, "r" becomes the implicit key. This conflicts with explicit key for the COM reg tool, so both should be removed in the mean time.