Honor useSpellingFunctionality parameter when reporting shortcuts (#15572)
Closes #15566
Fix-up of #14900 / #15373.

Summary of the issue:
In #14900 / #15373., spelling mode was used (CharacterModeCommand) no matter the state of the "Use spelling functionality" option.

The use spelling functionality exists because some old SAPI synth do not support spelling mode correctly. For them, this option needs to be honored.

Description of user facing changes
Use spelling functionality option will now be honored when reporting shortcuts.

Description of development approach
See logic in the code.
If use spelling functionality option is disabled, the key is just reported as is.