Default gesture for the screen curtain (#15590)
Closes #10560

Summary of the issue:
The screen curtain is missing a default gesture (working out of the box). Users want to have a default gesture in order not to have to define one themselves.

Description of user facing changes
NVDA+control+escape is defined as default gesture.

The gesture has been discussed in #10560:

The main key is quite easy to find on a keyboard; e.g. PrintScreen is much harder because it's not always at the same place or sometimes behind Fn key.
Usage of punctuation key for main key should be avoided because it differs from a local keyboard layout to another. Even if it may be translated by translators, the issue remains for people switching their keyboard layout. Punctuation keys are better suited for personal gestures.
shortcut using 3 keys and not two due to the probable infrequency of use with respect to other gestures/commands
with this gesture, there is no risk to activate the screen curtain by accident.

Description of development approach
Added in script's decorator and updated the documentation.