Reduce verbosity when reporting Excel cell formatting (#15560)
When requesting for cell formatting information in Excel (NVDA+F), there is over-verbose information:

"no border lines" is often reported, since Excel files with default borders are quite common
"background pattern none" is almost always reported since cells with other background patterns are very rare
Description of user facing changes
When formatting information is requested with NVDA+F, we will not report anymore a lack of specific formatting for border lines or for background pattern. This formatting will still be reported during navigation if the corresponding option is enabled in document formatting settings (respectively "cell borders" and "color").

In that sense, we follow the logic used for other formatting attributes that have a negative state (lack of such formatting) such as bold/italic/etc. E.g.: "no bold" is not formatted upon NVDA+F request, but it is still reported while navigating from bold to no bold if the corresponding doc formatting option is enabled.

Description of development approach
Do not provide info on background pattern in formatting info fields and changed the logic in speech.py to handle this.

For UIA, nothing is done since NVDA does not report borders or background pattern for Excel cells when UIA is enabled.