Move getFormattedStacksForAllThreads from watchdog to logHandler, to fix build of developer documentation. (#15616)
Related to #12971 problem caused by PR #14899

Summary of the issue:
When trying to build developer documentation with Sphinx, many modules failed to be parsed

Description of development approach
getFormattedStacksForAllThreads is moved from watchdog to the logHandler, which avoids these warnings. Appropriate backwards compatibility is added to watchdog, to make sure add-on authors are warned to use the moved function from logHandler in the future.