Move the column Status in the Add-on Store GUI (#15621)
Fixes #15069

Summary of the issue:
In the Add-on Store GUI the Status column is the last one. However, the status information is much more important for users than Author. Users wanting to hear the status information when an add-on is selected in the GUI need to wait for the whole line to be read or to use additional object navigation commands. This is less efficient.

Description of user facing changes
The column "Status" is now located just after the add-on's name.

As discussed in #15069, both version numbers and status are important informations in the Updatable add-on tab. However, it's important to keep a quite similar order between tabs to allow a smoother UX.

Description of development approach
In AddonListField enum, moved the status item in second position.