Remove deprecated speechDictHandler.speechDictVars. (#15629)
Closes #15614

Summary of the issue:
speechDictHandler.speechDictVars was marked as deprecated in PR #15021. Unfortunately since the module was not imported anywhere in core py2exe failed to bundle it, which made it impossible to import this code in binary versions of NVDA. Additionally warning which was raised when importing speechDictHandler.speechDictVars was problematic when building developer documentation. Since this API breakage was already included in two releases of NVDA, it was decided to remove the code and document the removal.

Description of user facing changes
None - this only removes deprecated code.

Description of development approach
The deprecated code is removed. Change log is updated to document the removal for add-ons developers.