Document context help in User Guide and Commands Quick reference (#15625)
Closes #13695

Summary of the issue:
Context help is not well known by users. It has been annonced 2-3 year ago in the Change Log, but it's a long time, now.

Description of user facing changes
Document context help in User Guide and in the Commands Quick Reference so that it is more widely known.