Indicate that setting column/row headers in Word tables does not work for UIA (#15617)
Fixes #13006

Summary of the issue:
UIA has become the default in Word for people with recent versions of Word and running Windows 11. When they try to set the column/row headers in tables with NVDA+shift+C/R, the command does not work and they are confused.
The first row and the first column are however automatically treated as headers.

Description of user facing changes
Following the discussion in #13006, ^more specifically #13006 (comment), and considering that:

the headers are usually in the first row or column in of the tables,
Some work is needed on Microsoft's side to make our traditional headers work with UIA,
it has been decided to accept and document this limitation of being able to have headers only on the first row/column.
Thus the User Guide has been updated.

Also, in Word documents when using UIA, NVDA+shift+C/R now report an information so that user know that these commands are not supported nor needed anymore.

Notes:
When this fix is released in a stable release, we can evaluate if this information is enough for people or if they keep on asking the old way to manage column/row headers; we can then consider reopening #13006 (or opening a new similar issue) if needed.

Description of development approach
Two script added in the class for Word UI§A documents. These script have no description since they do not need to be discovered nor remapped in the Input gesture dialog. They're just here to give information to people trying to use the old way to handle row/column headers.