Remove deprecated app module aliases. (#15639)
Closes #15618

Summary of the issue:
PR #13366 changed a way in which NVDA maps an app module to a given executable. In the process several old app modules were marked as deprecated (they show warnings on import). This is pretty noticeable when building developer documentation scons devDocs.

Description of user facing changes
None

Description of development approach
Deprecated app modules and supporting functions from appModuleHandler were removed. The app Module for Azardi which cannot be imported by Sphinx was renamed, and mapped to Azardi's binary.