winVersion/full screen magnification: say 'Windows 8.1'. Re #15664.
Mostly for screen curtain: as NVDA requires Windows 8.1 or later, there is no need to actualy provide a flag that says full screen magnification API is present. However the API is not supported for 32-bit apps (WWoW64), so keep the flag around in case Microsoft removes in a future Windows release.