Fix Configuration Profiles dialog on 3.11 (#15660)
Fixes #15650

Summary of the issue:
The profiles dialog broke after the upgrade to 3.11 do to a change in WxPython that made proportion argument required to be an int. Something similar occurred with the Vision Settings Panel when trying to enable the highlighter

Description of user facing changes
Profiles dialog works as expected and can be navigated and used normally without errors. As well as enabling the highlighter in settings.

Description of development approach
change the proportion argument to 1 instead of 1.0.