When logging role and states for dev info use names of the enum members. (#15708)
Fixes #15703

Summary of the issue:
In Alphas of NVDA compiled with Python 3.11 role and states in the developer info were shown as integer values of the underlying enum members, and not as, more readable, member names. This is due to a change in Python 3.11, which adds a new base class for IntEnum and IntFlag which converts the members to string in the same way their underlying type (in this particular case integers) are formatted. Since we convert both role and states implicitly (using %s) to string in the developer info the member description is replaced with the meaningless integer value there.

Description of user facing changes
Role and states are once again displayed as names in the developer info.

Description of development approach
When logging these properties we are logging their names directly, rather than converting enum members to string.