Keep a backup of a corrupted config when reinitializing it (#15779)
Related to #14424 and #14412

Summary of the issue:
In some situation, the configuration file nvda.ini may be corrupted. This can happen if the user has edited the file manually or in some cases when NVDA is downgraded. It may also happen in the case of a bug during the development phase of NVDA (alpha releases) or of add-ons.

When the configuration file is corrupted, it cannot be used. Thus NVDA deletes it and starts with a brand new config. It also warns the user with a message box indicating that the config has been reset and that more information can be found in the log. However in some case, the user may want to fix the issues in the corrupted file manually instead of re-creating all the parameters from the new config.

Description of user facing changes
Before re-creating a new config, a backup of the old corrupted file is saved so that the user can turn back to it if they want. The logged message also indicates the presence of this file.