Use UIA for syslistview32 controls backed by Windows Forms (#15799)
Fixes #15283

Summary of the issue:
In Windows Forms applications, errors occur in list view controls.

Description of user facing changes
Win forms project in #15283 now works as expected for the most part.

Description of development approach
Rather than the previous approach where we removed the listing of SysListView32 from the list of bad UIA class names, this adds a specific utility function to checks for a Framework Id of WinForm. If that's the case, we use UIA for SysListView32, as it seems these implementations are mature enough.