Mark add-on store code as public API (#15777)
Initially, the majority of new add-on store code was marked as part of the private API. This was because it was an extremely large change and there were expected and unexpected changes to happen to the code in subsequent releases.

The add-on store was introduced in 2023.2. By the time 2024.1 is released, we can expect the add-on store API to be relatively stable.