Report unavailable (disabled) objects during object navigation (#15785)
Closes #15477

Summary of the issue:
In a GUI, disabled (greyed) items are visible to sighted users. So unless there is a good reason, blind people should be able to see them too. Greyed items convey meaning in a GUI and not being able to see them causes people to miss information conveyed by the GUI.

For example in the new audio panel, if you run NVDA from source (or probably portable), the audio ducking item is greyed out (disabled). Being allowed to find the disabled item informs the user that NVDA has an audio ducking options but that this options can not be enabled.

Unfortunately, disabled controls are ignored during object navigation.

This is also inconsistant with document review mode where disabled objects are not ignored.

Description of user facing changes
During object navigation (complete, simple or flattened), disabled object will not be ignored anymore.

Description of development approach
For presentation type, do not classify objects with state UNAVAILABLE to presType_unavailable.
For windows isUsableWindow does not return False for windows that are not enabled