Allow more batch actions in the add-on store (remove, enable and disable) (#15646)
Closes #15623

Summary of the issue:
In the store, install action can be executed on multiple add-ons. Other batch actions are needed.

Description of user facing changes
The following additional batch actions are now available in the store:

Remove selected add-ons
Update selected add-ons
Enable selected add-ons
Disable selected add-ons

Description of development approach
Use the same condition as the one required for the corresponding single action to occur.
For code simplicity there is no distinction between Enable and Enable (override) for the corresponding multiple action. Anyway, when an add-on should override compatibility to be enabled, there is a confirmation dialog box.
Also for simplicity, there is no multiple replace action:
if a multiple selection contains installable add-on with or without updatable or replaceable add-ons, the "Install selected add-ons" batch action will be available
If a multiple selection contains updatable add-ons, replaceable add-ons or both types, the "Update selected add-ons" batch action will be available.
For batch actions requiring a confirmation (e.g. override incompatibility, uninstall), the confirmation dialog box has a "Remember this choice" checkbox so that the user can answer yes or no for all the same questions.
When many add-ons are selected and a batch action is applied to them, the action actually applies only to the add-ons for which it is appropriate. E.g batch enable will not be applied to already enabled add-ons.