Report keyboard shortcut when entering a list (#15826)
Closes #15816

Summary of the issue:
When the focus enters in a list, the list's keyboard shortcut is not reported. It's because a list item takes the focus, whereas the shortcut key is on the whole list; logically, individual list items have no shortcut. We usually do not report shortcut keys when entering container objects. However, in the case of list, it would be useful so that people know that the list has a shortcut key.

Description of user facing changes
Reports shortcut when the focus enters a list.

Description of development approach
When speaking an object for FOCUS_ENTERED reason, if the object is a list, we do not discard keyboard shortcut anymore.

While at it, removed the accelerator of path selection in the "Create portable NVDA" dialog, since I have realized during the investigation for this PR that it was not working. A better solution would be to have a working key accelerator for this field but I do not know how to do this