When installing add-ons in add-on store make sure that restart dialog is shown only when the install is done (#15852)
Fixes #15613

Summary of the issue:
When installing add-ons developers may want to present a GUI prompt as part of the install tasks. When installing from the external bundle this prompt is blocking, in other words user has to answer it before installation is complete and NVDA asks to be restarted. In comparison installations from the add-on store shows the prompt, but then the restart prompt was shown, making it impossible to interact with the add-ons installation GUI.

Description of user facing changes
When installing add-ons from the store the restart message is shown only when all installations are done, giving user a chance to interact with eventual GUI prompts they raise.

Description of development approach
Similar to the code path used when installing external bundles, call to the addonHandler.installAddonBundle is executed in a separate, blocking threat using ExecAndPump. ExecAndPump was also moved from gui to systemUtils, as it has no relation to GUI, other than the fact that it was used only in this package.